requirements.txt
python-dotenv
pandas
numpy
seaborn
matplotlib
pymongo
certifi
pymongo[srv]==3.11
scikit-learn

-e .